+++ pkg/local_object_storage/shard/control.go
@@ -255,15 +255,20 @@ func (s *Shard) Close() error {
 
 	components = append(components, s.blobStor, s.metaBase)
 
+	var lastErr error
 	for _, component := range components {
 		if err := component.Close(); err != nil {
-			return fmt.Errorf("could not close %s: %w", component, err)
+			lastErr = err
+			s.log.Error("could not close shard component", zap.Error(err))
 		}
 	}
 
-	s.gc.stop()
+	// If Init/Open was unsuccessful gc can be nil.
+	if s.gc != nil {
+		s.gc.stop()
+	}
 
-	return nil
+	return lastErr
 }
 
 // Reload reloads configuration portions that are necessary.